Move RequestStreamAsyncResult to its own file.
